stats: clear hot-restart parent contributions from child gauges when parent terminates. (#11301)
Commit Message: A typical gauge tracks some in-progress count within an Envoy process, which is expected to rise and eventually fall to zero. During the hot-restart process, the child stats for Gauges with an ImportMode of Accumulate include the parent data. However, not all parent gauges get zeroed and clean out their contributions to the child gauges, so do so forcibly on parent termination.

One exception to this rule is server.hot_restart_generation which should accumulate over the generations, by definition, so a mechanism was added to allow retention of accumulated gauges. It's possible others may need this as well. Arguably server.hot_restart_generation should be a counter and not a gauge, but due to the way counters transmit latched values over the generations, this will not work as is.
Additional Description:
Risk Level: medium -- will change behavior of some gauges.
Testing: //test/...
Docs Changes: n/a
Release Notes: n/a
Fixes #10806

Signed-off-by: Joshua Marantz <jmarantz@google.com>